New tool: getSessionState Clarified waitForBreakpoint
